requirements.txt
asgiref
attrs
beautifulsoup4
bs4
certifi
charset-normalizer
Django
django-redis
djangorestframework
h11
idna
outcome
packaging
pillow
PySocks
python-dotenv
redis
requests
selenium
sniffio
sortedcontainers
soupsieve
sqlparse
trio
trio-websocket
typing_extensions
urllib3
webdriver-manager
websocket-client
wsproto
